[nan-1762] updates to notion
