+++ integrations/notion/nango.yaml
@@ -1,33 +1,49 @@
 integrations:
     notion:
         syncs:
-            pages:
-                runs: every 6 hours
-                description: |
-                    Sync pages, subpages, database entries, entries of
-                    sub-databases and retrieve only the text content and ignores
-                    images, files and other blocks that do not have
-                    a `rich_text` property.
-                output: NotionPage
-                sync_type: incremental
-                endpoint: GET /notion/pages
-            rich-pages:
-                runs: every 6 hours
+            content-metadata:
+                runs: every 1h
                 track_deletes: true
                 description: |
-                    Sync pages, subpages, database entries, entries of
-                    sub-databases and maps to full markdown. It transforms
-                    images, tables, uploaded files, etc into their markdown counterparts.
-                output: NotionRichPage
-                sync_type: incremental
-                endpoint: GET /notion/rich-pages
+                    Sync pages and databases metadata to further fetch the content
+                    using a dedicated action
+                output: ContentMetadata
+                sync_type: full
+                endpoint: GET /metadata
+        actions:
+            fetch-rich-page:
+                description: |
+                    Fetch a specific page in Notion by passing a pageId. This action fetches a page,
+                    and its content and converts it into a full markdown. It transforms images,
+                    tables, uploaded files, etc., into their markdown counterparts, providing a complete markdown.
+                input: RichPageInput
+                output: RichPage
+                endpoint: GET /fetch-rich-page
+            fetch-database:
+                description: |
+                    Fetch a specific Notion database by passing in the database id. This action fetches
+                    the database and converts it into a CSV string.
+                input: DatabaseInput
+                output: Database
+                endpoint: GET /fetch-database
+            fetch-content-metadata:
+                description: |
+                    Retrieve the entity type as well as an id for a Notion entity to later call
+                    fetch-database or fetch-rich-page based on the type.
+                input: UrlOrId
+                output: ContentMetadata
+                endpoint: GET /fetch-content-metadata
 models:
-    NotionPage:
+    RichPageInput:
+        pageId: string
+    ContentMetadata:
         id: string
-        url: string
-        content: string
-        parent_page_id: string | undefined
-    NotionRichPage:
+        path?: string
+        type: page | database
+        last_modified: string
+        title?: string
+        parent_id?: string | undefined
+    RichPage:
         id: string
         path: string
         title: string
@@ -35,3 +51,18 @@ models:
         contentType: string
         meta: object
         last_modified: string
+        parent_id?: string | undefined
+    DatabaseInput:
+        databaseId: string
+    DatabaseEntry:
+        __string: string
+    Database:
+        id: string
+        path: string
+        title: string
+        meta: object
+        last_modified: string
+        entries: DatabaseEntry[]
+    UrlOrId:
+        url?: string
+        id?: string
