+++ package.json
@@ -9,6 +9,7 @@
         "move:integrations": "bash scripts/move-all-to-nango-integration-directories.bash $npm_config_integration",
         "undo:move:integrations": "bash scripts/undo-move-to-nango-directories.bash",
         "lint-moved-integrations": "npm run move:integrations && npm run lint && npm run undo:move:integrations",
+        "generate:zod": "bash scripts/generate-integration-template-zod.bash $npm_config_integration && npm run undo:move:integrations",
         "compile": "bash scripts/compile-all-templates.bash && npm run undo:move:integrations",
         "prettier-format": "prettier --config .prettierrc \"./**/*.{ts,tsx}\" --write",
         "prettier-watch": "onchange './**/*.{ts,tsx}' -- prettier --write {{changed}}",
@@ -33,9 +34,11 @@
         "lint-staged": "15.2.9",
         "nango": "^0.42.13",
         "onchange": "7.1.0",
+        "ts-to-zod": "^3.9.1",
         "tsx": "^4.19.0",
         "typescript": "5.3.3",
-        "vitest": "^2.0.5"
+        "vitest": "^2.0.5",
+        "zod": "3.23.8"
     },
     "lint-staged": {
         "*.{html,css,json,yaml}": [
